# Fixed items total to display total items, it was wrongly displaying CURRENT page items. Also enhanced it to display the current page items e.g. 31-40

+++ com_flexicontent_v2.x/admin/models/templates.php
@@ -122,7 +122,7 @@ class FlexicontentModelTemplates extends JModel
 
 		$path 	= JPATH_COMPONENT_SITE . DS . 'templates' . DS;
 		
-		if (!$dir || ($dir == 'blog') || ($dir == 'default')) return false;		
+		if (!$dir || ($dir == 'blog') || ($dir == 'default') || ($dir == 'faq') || ($dir == 'presentation')) return false;		
 		if (!JFolder::delete($path.$dir)) return false;
 		
 		// delete old record
